Convex hull points

Computes points to plot a convex hull for each cluster of the clustering model

convex.hulls(clustering.model, data.pca)

Arguments

clustering.model
Object returned by the cl.analysis function.

data.pca
data as result of the stats::prcomp(clustering.model$data, scale. = T, center = T).

Value

List of vectors with points for each convex hull.