Read and Calculate the Basic Information for Cluster Analysis Tasks from Open Spending API

Extract and analyze the input data provided from Open Spending API, using the cl.analysis function.

open_spending.cl(json_data, dimensions=NULL, amounts=NULL, measured.dimensions=NULL,
cl.aggregate="sum", cl.method=NULL, cl.num=NULL, cl.dist="euclidean")

Arguments

json_data
The json string, URL or file from Open Spending API

dimensions
The dimensions/feature of the input data

amounts
The measures of the input data

measured.dimensions
The dimensions to which correspond amount/numeric variables

cl.aggregate
Aggregate function of the input data

cl.method
The clustering algorithm

cl.num
The number of clusters

cl.dist
The distance metric

Value

A json string with the resulted parameters of the cl.analysis function.

Details

This function is used to read data in json format from Open Spending API, in order to implement cluster analysis through cl.analysis function.

See also

cl.analysis