Sample city data

This dataset is an example data frame of the budget phase data

Administrative_Unit

Approved

Draft

Executed

Revised

Format

A data frame with the previous characteristics as columns