Clustering features

Select clustering characteristic to form the clustering data

cl.features(data, features = NULL, amounts = NULL, aggregate = "sum", tojson = FALSE )

Arguments

data
The input data

features
The clustering features

amounts
The amount measures of the dataset

aggregate
The function to aggregate

tojson
If TRUE the results are returned in json format, default returns a list

Value

This function returns the dataset for cluster analysis adapted to the desired features.

Details

This function adapts the dataset according to the selected dimension of the dataset and the aggregation function.

See also

cl.analysis